requirements.txt
Django>=4.2,<4.3
django-environ>=0.11
Pillow>=10.0
gunicorn>=21.2
whitenoise>=6.6.0
